+++ R/data.R
@@ -52,6 +52,32 @@
 #' table(proms$which.tissues)
 "proms"
 
+#' ce11_TSSs
+#'
+#' Annotated TSSs in C. elegans (ce11) according to Serizay et 
+#' al. 2020, "Tissue-specific profiling reveals distinctive regulatory 
+#' architectures for ubiquitous, germline and somatic genes", BiorXiv.
+#'
+#' @docType data
+#'
+#' @usage data(proms)
+#'
+#' @format GRanges
+#'
+#' @keywords datasets
+#'
+#' @references Serizay et al. 2020, 
+#' "Tissue-specific profiling reveals distinctive regulatory 
+#' architectures for ubiquitous, germline and somatic genes", BiorXiv.
+#' (\href{https://doi.org/10.1101/2020.02.20.958579}{DOI})
+#'
+#' @source \href{https://doi.org/10.1101/2020.02.20.958579}{BiorXiv}
+#'
+#' @examples
+#' data(ce11_TSSs)
+#' lengths(ce11_TSSs)
+"ce11_TSSs"
+
 #' ce_proms_seqs
 #'
 #' Sample of sequences of promoters annotated in C. elegans (ce11) according to